Remove indentation assumption in RTP parser
Charmm27 RTP file has an unindented '[cmap]' subsection. This shows
that, even though the indentation of subsection is usually present it is
only a convention (or there is a bug in gromacs...).

This commit replaces the use of the indentation to recognize subsection
by an exhaustive list of the subsection names.