Actually read the bondedtypes section of RTP files
The bondedtypes section of a RTP file describes what is the function
type of the different interactions, and how to handle automatically
generated interactions.

This commit add the parsing of the bondedtypes section, the generation
of the dihedral angles according to the rules in that section, and the
assignation of the function types.

Note that the selection of the dihedral angles to keep may not yet be
correct, and that the 1-4 pair interactions between hydrogen atoms is not
implemented.