Add a ForceField class
The ForceField class exposes force field dependant data in a normalize
way so processors do not need to know what force field they work on.

For now, a force field object exposes the blocks, links, and reference
graphs.

The force field is accessible *via* the Molecule and the System objects
with the 'force_field' property. It can be assigned for System object by
assigning a force field to the 'force_field' property, then the force
field is propagated to all the molecules of that system. It is
recommended not to change the force field at the Molecule level as it
could break the consistency between molecules in a system. However, it
is possible to reassign the '_force_field' attribute of Molecule object,
or to pass the 'force_field' argument when creating a new Molecule
instance.

The force field class is very primitive yet. An instance is built from a
directory that contains one or more RTP files. The name of the directory
is used as name for the force field.

Using the class allows to get rid of the hardcoded RTP path in
apply_blocks and do_links. The hardcoded path is still present in
do_mapping. (See #4).