Adds a DSSP processor
This commit introduces the dssp module that allows to run DSSP and to
annotate a system with the secondary structure assignations.

To reach this goal, the commit also introduces the possibility to
generate a PDB file as a string. This change assumes that a PDB file
will always fit in memory as, now, the whole string is generated before
being written on file.

I only tested DSSP on a crystal structure, not yet on a NMR structure
with multiple models.

This is a first step toward secondary-structure-dependant topologies.
See #22.