Interactions can be written grouped in ITP files
The order of the interactions in an ITP file have a great effect on the
readability of the file. This commit allow to group interactions within
a section when they are semantically related. For instance, it allows to
group all the backbone-backbone bonds when dealing with a protein.

Interactions are grouped on the basis of their "group" meta attribute.
This attributes takes a human readable string that will be written as a
comment at the beginning of the group. Interactions without a "group"
meta attribute are written first.