Link can exclude an edge
It is possible to define an edge that must be absent for the link to be
applied. This allows, for instance, to define links at the extremities
of a polymer.

The edges that must non exist are listed in the `non_edges` attribute of
the `Link` object as pairs of the name of a node that exists in the
link, and the attributes of the node it must not be attached to.