Link replace interactions instead of cumulating them
If two links define the same interaction, the later one replaces the one
defined by the previous. This is done using the new
`Molecule.add_or_replace_interaction`.

In the case a user explicitly wants to define the same interaction
multiple time, he can tag them with the attribute "version". This
attribute is assumed to be 0 if not defined. Two interactions are the
same if they are in the same section, target the same nodes in the same
order, and have the same "version".