Add support for dummy charge particles
Martini 2.2P uses dummy beads to hold the charges. These beads are not
part of the mapping; therefore, they must be added to the mapping target
molecule without having underlying atoms, and there coordinates must be
constructed.

This commit:
* Adds an [ extra ] section in the mapping files to declare unmapped particles
* Adapt do_mapping to add the unmapped particles
* Adds the LocateChargeDummies that construct the coordinates of the
  dummies
* Adds the force field file for martini 2.2P

In addition, I added a [ from ] and a [ to ] sections to the mapping
files as a quick fix to specify what force field the files are for.

The AverageBeads processor can ignore beads that do not have an
underlying graph.