Prevent attributes from AA to pollute CG
During the mapping step, attributes from the original molecules are
passed on to the mapped ones. This leads to the mapped molecules to have
wrong attributes, such as atomistic masses on a coarse grained bead.

This commit allows to keep only the attributes listed as argument. The
attributes required to identify a residue are passed on in any case.
These default attributes are chain, resid, and resname.