Make PDB writer more robust
Some node attributes may be set to None when they are unknown, instead
of being undefined. This means that `node.get(attr, 0)` can return
`None` even if there is no known valid value. This breaks the PDB writer
is some situations.

This commit makes sure that the PDB writer uses its default value wjen
an attribute is set to `None`.