More options for links residue order
The order can now be a series of + (i.e. +, ++, +++, ...) to indicate
that the residue has any resid greater than the reference, a series of -
to indicate that the resid is lesser, or a series of * to indicate any
other resid than the reference.

The full comparison matrix between the different possible values of
order is described in the docstring of
vermouth.processors.do_links._match_order.