Fix the -write-* debug options in martinize2
The -write-* options allow to write the state of the system as a PDB
file at various stages of the workflow. Though, the wrong system was
wtitten. Also, some stages lack coordinates for some atoms which was
causing issues. This commit adds the `omit_missing_pos` argument to the
PDB writer that allows to write nan when the coordinates are missing.