Let scipy calculate the distances in make_bonds
make_bonds was using a KDTree to find the pairs close enough to be
considered, and then recalculated the distances for these pairs one by
one. This commit uses KDTree.sparse_distance_matrix to get both the
pairs and the distances from scipy. In addition to other minor
optimisations in the function, the performances are ~10 times better for
a given example.

Because the redistributed KDTree does not provide the
sparse_distance_matrix method, this commit implements a substitute.