Faster edge selection in Molecule.subgraph
The edge selection was the slower part of Molecule.subgraph. This
performance hit was affecting heavily do_mapping that creates a subgraph
for each bead at the final resolution.

The performance cost was due to the iteration over the edges of a graph
being slow. This commit instead iterate over the neighbors of the nodes
in the selection.