Make the scFix optional
This commit introduces the -scfix option to the command line. This
option sets the value of the molecule meta "scfix" to `True` and expects
the "scfix" feature to be implemented by the force field.

The commit also make the scFix related links of martini30 depends on the
"scfix" molmeta so they are controlled by the option.

Fixes #123