Tests for the force field class and the force field parser (#102)
* Reduce the need to load all the force fields

We do not systematically load all the native force fields when we import
vermouth.forcefield. Instead we load them all only if we need one of
them. It is not an ideal solution, but it will reduce the test coverage
noise when running individual test cases. The noise is still there when
running the full test suite as some tests do request for a specific
native force field, therefore triggering the force field parsing
mechanism.

* Add tests for the ForceField class

* Add some tests for the ff parser

* Fix compatibility issue

* Moar tests for ffinput

* Remove the last reference of forcefield.FORCE_FIELDS

* Address @pckroon comments on #102

* New tests for the ff parser and random bug fixing

* Fix optional keywords in sphinx

The numpy docstring format says that optional arguments should be marked
with the "optional" keyword. Though, this displease the nitpicky mode of
sphinx that complains about not finding the corresponding reference.

This commit makes sphinx ignore that specific warning.