Add a processor that assign moltype names to molecules (#184)
* Quick fix moltype deduplication

After #103, Molecule became unhashable which broke the deduplication of
moltypes in martinize2. This commit avoids the need for Molecule to be
hashable by avoiding the use of a dict.

* Processor to assign moltypes

* Replace martinize2 moltype assignation by the processor

* Fix test_itp after remodeling moltype

With the moltype saved as a meta attribute rather than as an attribute,
the dummy molecule in test_itp had to be adapted.

* Move molecule hypothesis strategies in their own module

* Add tests for NameMolType

* Document tests for NameMolType

* Have martinize2 called in the tests to validate NameMolTypes

* Improve performances in test_name_moltype

* Address @pckroon's comments on #184

* Fix an wrong reference in the doc

* Please pylint

* Run pytest from coverage rather than the opposite

Calling subprocess in the tests makes coverage ignore its configuration
when called from pytest. Coverage appears to work as expected when
pytest is run from coverage.

When ignoring its configuration, coverage does not omit the tests and
the redistributed files from its report.

* Hopefully avoid test_link_equal to timeout

* Typos

* Fix a mistake reporting coverage in travis