Allows to write arbitrary lines at the beginning or end of ITP sections
`write_molecule_itp` now has a `pre_section_lines` and a
`post_section_lines` that describe lines to write just after the section
header, or right after the last line of the section, respectively.

If the arguments are not provided, the function will try to read the
line descriptions from `molecule.meta`.

One use for the feature is to add include statements.