Move sorting atoms into a processors
Atoms were sorted at the time of writing the output PDB or GRO file.
This sorting is required to keep reconstructed atoms in their respective
residues. However, the code was redundant between the PDB write function
and the GRO one, and it was not done in the ITP writing function which
could lead to inconsistencies. Finally, the sorting could not be
bypassed.

This commit adds a sorting processor. The sorting can therefore be done
after the grap is repaired and canonicalized.