Add somme tests for apply_rubber_band
Some of the tests are failing. They do represent the desired behaviour
of apply_rubber_band.build_connectivity_matrix rather than the actual
one.

At the moment, apply_rubber_band.build_connectivity_matrix builds a
subgraph with the selected nodes and get the connectivity matrix from
there. However, it means that a selection that would only include the CA
atoms/beads of an aomistic or sirah protein could see two consecutive
CA particles as not connected even if the separation is large enough
that they should appear connected. The tests assert for the correct
behaviour and assume the function only builds the matrix for the
selected beads, but uses the full graph for the connectivity.