Fix application of modifications via CLI that do not result in PTM atoms
The data flow was as follows:
1) CLI
2) annotate mutmod (annotates)
3) repair graph (repairs with modified reference). PTM_atom attributes
from the modification are added automatically for the relevant nodes.
4) canonicalize modifications figures out that some atoms are not part
of a block, and identifies the corresponding modification.
5) do mapping looks at the 'modifications' attribute set by canon mod to
figure out the mod mappings.

Now, if a requested modification does not add any PTM atom, no atoms
will get marked, which means the canonicalize modifications processor
does nothing, which in turn means the modification does not get mapped
to the output files.

This commit fixes that: repair graph now marks the modification touched
atoms with a 'modifications' attribute, which then gets used by canon
mod.