Move martinize2 into a vermouth function
Bash command martinize2 is now a function in vermouth.martinize2.
The bash command directly refers to it, so only one function needs
to be modified.

One issue for the lazy or forgetful: arguments and their documentation
now need to be modified two times (in the CLI and the python API).
This should however make martinize2 more easily testable.

Type hints could be added to the martinize2 function.