Answer Sphinx's three riddles to appease it
Adapt the martinize2 docstring so that sphinx/napoleon can work
with it. In essence, wrap lists of options and defaults in
backticks. Remove a lot of defaults from the docstring, since they
are the same as in the signature.

Also remove the parameter subsections. It /is/ possible to add
those to napoleon (napoleon_custom_sections in conf.py), but they're
not as pretty as the normal parameters section, plus, it bloats the
conf.py